refactored cmd/suti.go to work with new API
- Data is now defined here
- added loadFilePaths for parsing paths as a list of filepaths
- added mergeData
- suti_test now run build before running suti to catch compile errors

+++ cmd/suti_test.sh
@@ -1,12 +1,20 @@
 #!/bin/bash
 # if nothing prints, the test passed
 
-diff="diff"
+diff="diff -bs"
 
-go run suti.go -cfg ../examples/suti.cfg -r ../examples/template/html.hmpl > out.html
-$diff out.html ../examples/out.html
-rm out.html
+go build -o suti suti.go
 
-go run suti.go -cfg ../examples/suti.cfg -r ../examples/template/txt.mst > out.txt
-$diff out.txt ../examples/out.txt
-rm out.txt
+if [ -e suti ]; then
+	./suti -cfg ../examples/suti.cfg -r ../examples/template/html.hmpl > out.html
+	$diff out.html ../examples/out.html
+	rm out.html
+
+	./suti -cfg ../examples/suti.cfg -r ../examples/template/txt.mst > out.txt
+	$diff out.txt ../examples/out.txt
+	rm out.txt
+
+	rm suti
+
+	echo "if files are identical, TEST PASS"
+fi
